+++ resolwe_bio/tests/processes/test_enrichment.py
@@ -1,3 +1,6 @@
+from pathlib import Path
+
+from resolwe.flow.models import Data
 from resolwe.test import tag_process, with_resolwe_host
 
 from resolwe_bio.utils.test import KBBioProcessTestCase
@@ -7,12 +10,14 @@ class EnrichmentProcessorTestCase(KBBioProcessTestCase):
     @with_resolwe_host
     @tag_process("goenrichment")
     def test_go_enrichment_dicty(self):
+        input_folder = Path("goenrichment") / "input"
+        output_folder = Path("goenrichment") / "output"
         with self.preparation_stage():
-            inputs = {"src": "ontology_dicty_cropped.obo.gz"}
+            inputs = {"src": input_folder / "ontology_dicty_cropped.obo.gz"}
             ontology = self.run_process("upload-obo", inputs)
 
             inputs = {
-                "src": "gaf_dicty_cropped.gz",
+                "src": input_folder / "gaf_dicty_cropped.gz",
                 "source": "DICTYBASE",
                 "species": "Dictyostelium discoideum",
             }
@@ -33,22 +38,51 @@ class EnrichmentProcessorTestCase(KBBioProcessTestCase):
             enrichment,
             enrichment.output["terms"],
             "",
-            "go_enriched_terms_dicty.json.gz",
+            output_folder / "go_enriched_terms_dicty.json.gz",
         )
+
+        inputs = {
+            "ontology": ontology.pk,
+            "gaf": annotation.pk,
+            "pval_threshold": 1,
+            "source": "DICTYBASE",
+            "species": "Mus musculus",
+            "genes": ["DDB_G0277589", "DDB_G0286855", "DDB_G0267640"],
+        }
+        enrichment = self.run_process("goenrichment", inputs, Data.STATUS_ERROR)
+        warning_msg = [
+            "Selected genes Species must be the same as the Species "
+            "field of the GAF file."
+        ]
+        error_msg = [
+            "Selected genes are from Mus musculus, while GAF "
+            "file has defined Dictyostelium discoideum under Species field."
+        ]
+        self.assertEqual(enrichment.process_warning, warning_msg)
+        self.assertEqual(enrichment.process_error, error_msg)
 
     @with_resolwe_host
     @tag_process("goenrichment")
     def test_go_enrichment(self):
+        input_folder = Path("goenrichment") / "input"
+        output_folder = Path("goenrichment") / "output"
         with self.preparation_stage():
-            inputs = {"src": "ontology_mus_cropped.obo.gz"}
+            inputs = {"src": input_folder / "ontology_mus_cropped.obo.gz"}
             ontology = self.run_process("upload-obo", inputs)
 
             inputs = {
-                "src": "gaf_mgi_cropped.gz",
+                "src": input_folder / "gaf_mgi_cropped.gz",
                 "source": "MGI",
                 "species": "Mus musculus",
             }
             gaf = self.run_process("upload-gaf", inputs)
+
+            inputs = {
+                "src": input_folder / "gaf_mgi_cropped.gz",
+                "source": "AFFY",
+                "species": "Mus musculus",
+            }
+            gaf_error = self.run_process("upload-gaf", inputs)
 
         inputs = {
             "ontology": ontology.pk,
@@ -68,5 +102,31 @@ class EnrichmentProcessorTestCase(KBBioProcessTestCase):
             enrichment,
             enrichment.output["terms"],
             "",
-            "go_enriched_terms_mouse.json.gz",
+            output_folder / "go_enriched_terms_mouse.json.gz",
         )
+
+        inputs = {
+            "ontology": ontology.pk,
+            "gaf": gaf.pk,
+            "pval_threshold": 1,
+            "genes": ["GCTAG33G", "CGAGCG323"],
+            "source": "NCBI",
+            "species": "Mus musculus",
+        }
+
+        enrichment = self.run_process("goenrichment", inputs, Data.STATUS_ERROR)
+        error_msg = ["No genes were fetched from the knowledge base."]
+        self.assertEqual(enrichment.process_error, error_msg)
+
+        inputs = {
+            "ontology": ontology.pk,
+            "gaf": gaf_error.pk,
+            "pval_threshold": 1,
+            "genes": ["193202", "56535"],
+            "source": "NCBI",
+            "species": "Mus musculus",
+        }
+
+        enrichment = self.run_process("goenrichment", inputs, Data.STATUS_ERROR)
+        error_msg = ["Failed to map features."]
+        self.assertEqual(enrichment.process_error, error_msg)